requirements.txt
Flask
python-binance
pyTelegramBotAPI